Correction pour faire marcher le test d'un QCM avec enregistrement et affichage des réponses de l'utilisateur. Il reste à colorer les réponses de l'utilisateur en rouge et vert selon leur exactitude et à fermer la conversation.

+++ src/java/fr/rgrin/projetqcm/evaluation/EvaluateurNoteQcm.java
@@ -17,5 +17,5 @@ public interface EvaluateurNoteQcm {
    * les réponses de l'utilisateur.
    * @return la note
    */
-  public int calculNote(Questionnaire questionnaire);
+  public double calculNote(Questionnaire questionnaire);
 }
